+++ library/klip-api/src/jvmMain/kotlin/File.kt
@@ -9,3 +9,8 @@ public actual typealias File = java.io.File
 public actual fun File.writeText(text: String): Unit = kWriteText(text)
 public actual fun File.readText(): String = kReadText()
 public actual fun File.deleteRecursively(): Boolean = kDeleteRecursively()
+
+/**
+ * Native file separator for the platform (thanks a bunch, windows...)
+ */
+public actual val File.separator: String get() = File.pathSeparator