+++ library/klip-api/src/mingwMain/kotlin/File.kt
@@ -1,0 +1,6 @@
+package dev.petuska.klip.ext
+
+/**
+ * Native file separator for the platform (thanks a bunch, windows...)
+ */
+public actual val File.separator: String get() = "\\"