+++ plugin/klip-kotlin-plugin/src/main/kotlin/KlipCommandLineProcessor.kt
@@ -1,7 +1,8 @@
 package dev.petuska.klip.plugin
 
 import com.google.auto.service.AutoService
-import dev.petuska.klip.plugin.util.KlipKeys
+import dev.petuska.klip.plugin.config.KOTLIN_PLUGIN_ID
+import dev.petuska.klip.plugin.util.KlipOption
 import org.jetbrains.kotlin.compiler.plugin.AbstractCliOption
 import org.jetbrains.kotlin.compiler.plugin.CliOption
 import org.jetbrains.kotlin.compiler.plugin.CliOptionProcessingException
@@ -13,23 +14,20 @@ import org.jetbrains.kotlin.config.CompilerConfiguration
  */
 @AutoService(CommandLineProcessor::class)
 class KlipCommandLineProcessor : CommandLineProcessor {
-  override val pluginId: String = KlipMap.kotlinPluginId
+  override val pluginId: String = KOTLIN_PLUGIN_ID
 
   override val pluginOptions: Collection<CliOption> = setOf(
     CliOption(
       optionName = KlipOption.Enabled.name,
       valueDescription = KlipOption.Enabled.valueDescription,
       description = KlipOption.Enabled.description,
-    ),
-    CliOption(
-      optionName = KlipOption.Root.name,
-      valueDescription = KlipOption.Root.valueDescription,
-      description = KlipOption.Root.description,
+      required = true
     ),
     CliOption(
       optionName = KlipOption.Update.name,
       valueDescription = KlipOption.Update.valueDescription,
       description = KlipOption.Update.description,
+      required = true
     ),
     CliOption(
       optionName = KlipOption.KlipAnnotation.name,
@@ -53,11 +51,10 @@ class KlipCommandLineProcessor : CommandLineProcessor {
     configuration: CompilerConfiguration
   ) {
     when (option.optionName) {
-      KlipOption.Enabled.name -> configuration.put(KlipKeys.KEY_ENABLED, value.toBoolean())
-      KlipOption.Root.name -> configuration.put(KlipKeys.KEY_ROOT, value)
-      KlipOption.Update.name -> configuration.put(KlipKeys.KEY_UPDATE, value.toBoolean())
-      KlipOption.KlipAnnotation.name -> configuration.appendList(KlipKeys.KEY_KLIP_ANNOTATION, value)
-      KlipOption.ScopeAnnotation.name -> configuration.appendList(KlipKeys.KEY_SCOPE_ANNOTATION, value)
+      KlipOption.Enabled.name -> configuration.put(KlipOption.Enabled.key, value.toBoolean())
+      KlipOption.Update.name -> configuration.put(KlipOption.Update.key, value.toBoolean())
+      KlipOption.KlipAnnotation.name -> configuration.appendList(KlipOption.KlipAnnotation.key, value)
+      KlipOption.ScopeAnnotation.name -> configuration.appendList(KlipOption.ScopeAnnotation.key, value)
       else -> throw CliOptionProcessingException("Unknown option: ${option.optionName}")
     }
   }