+++ plugin/klip-kotlin-plugin/src/main/kotlin/KlipComponentRegistrar.kt
@@ -1,14 +1,16 @@
 package dev.petuska.klip.plugin
 
 import com.google.auto.service.AutoService
-import dev.petuska.klip.plugin.util.KlipKeys
+import dev.petuska.klip.plugin.util.KlipOption
+import dev.petuska.klip.plugin.util.KlipSettings
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.cli.common.CLIConfigurationKeys
 import org.jetbrains.kotlin.cli.common.messages.MessageCollector
+import org.jetbrains.kotlin.cli.js.messageCollectorLogger
 import org.jetbrains.kotlin.com.intellij.mock.MockProject
 import org.jetbrains.kotlin.compiler.plugin.ComponentRegistrar
 import org.jetbrains.kotlin.config.CompilerConfiguration
-import java.io.File
+import org.jetbrains.kotlin.name.FqName
 
 /**
  * Component Registrar responsible for validating command line inputs and registering required IR extensions
@@ -17,22 +19,16 @@ import java.io.File
 class KlipComponentRegistrar : ComponentRegistrar {
   override fun registerProjectComponents(project: MockProject, configuration: CompilerConfiguration) {
     val messageCollector = configuration.get(CLIConfigurationKeys.MESSAGE_COLLECTOR_KEY, MessageCollector.NONE)
-    if (configuration[KlipKeys.KEY_ENABLED] == false) {
-      return
-    }
-
     IrGenerationExtension.registerExtension(
       project,
       KlipIrGenerationExtension(
         settings = KlipSettings(
-          root = File(
-            configuration[KlipKeys.KEY_ROOT] ?: error("klip plugin requires source root option passed to it")
-          ),
-          update = configuration[KlipKeys.KEY_UPDATE] ?: KlipOption.Update.default,
-          klipAnnotations = configuration[KlipKeys.KEY_KLIP_ANNOTATION] ?: listOf(),
-          scopeAnnotations = configuration[KlipKeys.KEY_SCOPE_ANNOTATION] ?: listOf(),
+          enabled = configuration[KlipOption.Enabled.key] == true,
+          update = configuration[KlipOption.Update.key] == true,
+          klipAnnotations = configuration[KlipOption.KlipAnnotation.key]?.map { FqName(it) } ?: listOf(),
+          scopeAnnotations = configuration[KlipOption.ScopeAnnotation.key]?.map { FqName(it) } ?: listOf(),
         ),
-        messageCollector = messageCollector,
+        logger = messageCollectorLogger(messageCollector),
       )
     )
   }